[Bug] fixes the category and the recipes authentication and authorization.
This will fix the issue of getting and updating other peoples recipes.

[Fixes #154468192]

+++ classes/recipes.py
@@ -1,10 +1,11 @@
 # This file provides the CRUD operations for the recipes.
 
 from flask import request, jsonify, make_response
-from app.models import Recipes
+from app.models import Recipes, Categories
 from flask.views import MethodView
 from classes.auth.auth import token_required
 import re
+
 
 
 class NonFilteredRecipesManipulations(MethodView):
@@ -51,7 +52,7 @@ class NonFilteredRecipesManipulations(MethodView):
          422:
            description: Please fill all the fields
         """
-        recipe_name = str(request.data.get('recipe_name', '')).strip()
+        recipe_name = str(request.data.get('recipe_name', '')).strip().lower()
         recipe_procedure = str(request.data.get('recipe_procedure', '')).strip()
 
         check_recipe_existence = Recipes.query.filter_by(users_id=user_in_session,
@@ -120,54 +121,56 @@ class NonFilteredRecipesManipulations(MethodView):
              422:
                description: Please fill all the fields
                 """
-
         page_number = request.args.get("page", default=1, type=int)
         no_items_per_page = request.args.get("limit", default=10, type=int)
         search_recipe = request.args.get('q')
         # These are the arguments to be provided in the url
-
-        category_recipes = Recipes.get_all(category_id).paginate(page_number,
-                                                                 no_items_per_page, error_out=False)
+        check_category_exists = Categories.query.filter_by(id=category_id).first()
+        category_recipes = Recipes.query.filter(Recipes.category_id == category_id,
+                                                Recipes.users_id == user_in_session).\
+            paginate(page_number, no_items_per_page, error_out=True)
         # This retrieves all the recipes belonging to a specific category.
-
         recipe_list = []
-        if search_recipe:
-            # This filters the recipes when the user provides the search parameters q.
-            search_recipes = Recipes.query.filter(Recipes.users_id == user_in_session,
-                                                  Recipes.recipe_name.ilike('%' + search_recipe + '%')).\
-                paginate(page_number, no_items_per_page, error_out=False)
-            for recipes in search_recipes.items:
-                # This loops retrieves all the recipes that match the search string.
-                searched_recipes = {
-                    'id': recipes.id,
-                    'recipe_name': recipes.recipe_name,
-                    'recipe_description': recipes.recipe_description,
-                    'category_id': recipes.category_id,
-                    'date_created': recipes.created_at,
-                    'date_updated': recipes.updated_at
+        if check_category_exists:
+
+            if search_recipe:
+                # This filters the recipes when the user provides the search parameters q.
+                search_recipes = Recipes.query.filter(Recipes.users_id == user_in_session,
+                                                      Recipes.recipe_name.ilike('%' + search_recipe + '%')). \
+                    paginate(page_number, no_items_per_page, error_out=False)
+                for recipes in search_recipes.items:
+                    # This loops retrieves all the recipes that match the search string.
+                    searched_recipes = {
+                        'id': recipes.id,
+                        'recipe_name': recipes.recipe_name,
+                        'recipe_description': recipes.recipe_description,
+                        'category_id': recipes.category_id,
+                        'date_created': recipes.created_at,
+                        'date_updated': recipes.updated_at
                     }
-                recipe_list.append(searched_recipes)
+                    recipe_list.append(searched_recipes)
 
-            if len(recipe_list) <= 0:
-                # This checks whether the searched recipe list contains data.
-                return make_response(jsonify({'message': "Recipe does not exist."})), 404
-            return make_response(jsonify(recipe_list)), 200
+                if len(recipe_list) <= 0:
+                    # This checks whether the searched recipe list contains data.
+                    return make_response(jsonify({'message': "Recipe does not exist."})), 404
+                return make_response(jsonify(recipe_list)), 200
+            else:
+                for recipes in category_recipes.items:
+                    # This loops all the recipes of the category with no search parameters.
+                    all_recipes = {
+                        'id': recipes.id,
+                        'recipe_name': recipes.recipe_name,
+                        'recipe_description': recipes.recipe_description,
+                        'category_id': recipes.category_id,
+                        'date_created': recipes.created_at,
+                        'date_updated': recipes.updated_at
+                        }
+                    recipe_list.append(all_recipes)
+                if len(recipe_list) <= 0:
+                    return make_response(jsonify({'message': "No recipes for this category."})), 404
+                return make_response(jsonify(recipe_list)), 200
 
-        for recipes in category_recipes.items:
-            # This loops all the recipes of the category with no search parameters.
-            all_recipes = {
-                'id': recipes.id,
-                'recipe_name': recipes.recipe_name,
-                'recipe_description': recipes.recipe_description,
-                'category_id': recipes.category_id,
-                'date_created': recipes.created_at,
-                'date_updated': recipes.updated_at
-                 }
-            recipe_list.append(all_recipes)
-
-        if len(recipe_list) <= 0:
-            return make_response(jsonify({'message': "This category does not have such recipe."})), 401
-        return make_response(jsonify(recipe_list)), 200
+        return make_response(jsonify({'message': "Category does not exist."})), 404
 
 
 class FilteredRecipesManipulations(MethodView):
@@ -200,23 +203,24 @@ class FilteredRecipesManipulations(MethodView):
          404:
            description: Recipe/category not found
         """
-        single_recipe = Recipes.query.filter_by(users_id=user_in_session,
-                                                category_id=category_id,
-                                                id=recipe_id).first()
-        if single_recipe:
-            # This checks whether the provided recipe exists
-            one_recipe = {
-                'id': single_recipe.id,
-                'recipe_name': single_recipe.recipe_name,
-                'recipe_description': single_recipe.recipe_description,
-                'category_id': single_recipe.category_id,
-                'date_created': single_recipe.created_at,
-                'date_updated': single_recipe.updated_at
-            }
-            response = jsonify(one_recipe)
-            response.status_code = 200
-            return response
-        return make_response(jsonify({'message': 'Recipe does not exist'})), 404
+        check_category_exists = Categories.query.filter_by(id=category_id).first()
+        single_recipe = Recipes.query.filter_by(users_id=user_in_session, id=recipe_id).first()
+        if check_category_exists:
+            if single_recipe:
+                # This checks whether the provided recipe exists
+                one_recipe = {
+                    'id': single_recipe.id,
+                    'recipe_name': single_recipe.recipe_name,
+                    'recipe_description': single_recipe.recipe_description,
+                    'category_id': single_recipe.category_id,
+                    'date_created': single_recipe.created_at,
+                    'date_updated': single_recipe.updated_at
+                }
+                response = jsonify(one_recipe)
+                response.status_code = 200
+                return response
+            return make_response(jsonify({'message': 'Recipe does not exist'})), 404
+        return make_response(jsonify({'message': 'Category not found.'})), 404
 
     @classmethod
     def put(self, user_in_session, category_id, recipe_id):
@@ -262,12 +266,14 @@ class FilteredRecipesManipulations(MethodView):
              422:
                description: Please fill all the fields
         """
-        recipe_name = str(request.data.get('recipe_name', '')).strip()
+        recipe_name = str(request.data.get('recipe_name', '')).strip().lower()
         retrieve_recipe = Recipes.query.filter_by(users_id=user_in_session,
                                                   id=recipe_id,
                                                   category_id=category_id).first()
         # The retrieve_recipe stores the result of recipe to be updated.
-
+        check_category_exists = Categories.query.filter_by(id=category_id).first()
+        if not check_category_exists:
+            return make_response(jsonify({'message': 'Category not found.'})), 404
         if not retrieve_recipe:
             return make_response(jsonify({'message': 'Recipe does not exist.'})), 404
 
@@ -313,9 +319,11 @@ class FilteredRecipesManipulations(MethodView):
             401:
               description: Recipe not found
         """
-        recipe = Recipes.query.filter_by(users_id=user_in_session, id=recipe_id,
-                                         category_id=category_id).first()
+        recipe = Recipes.query.filter_by(users_id=user_in_session, id=recipe_id).first()
         # The recipe variable stores the result if the specific recipe to be deleted.
+        check_category_exists = Categories.query.filter_by(id=category_id).first()
+        if not check_category_exists:
+            return make_response(jsonify({'message': 'Category not found.'})), 404
         if recipe:
             recipe.delete()
             return make_response(jsonify({'message': 'Recipe deleted'})), 200
